HangulAutomata {KoNLP}	R Documentation

do Hangul automata

Description

function to be used for converting to complete Hangul syllables from Jamo or Keystrokes. Example will be shown in github wiki.

Usage

HangulAutomata(input, isKeystroke = F, isForceConv = F)

Arguments

input
to be processed mostly Jamo sequences

isKeystroke
boolean parameter to check input is keystroke or Jamo sequences

isForceConv
boolean parameter to force converting if input is not valid Jamo or keystroke sequences.

Value

complete Hangul syllable

[Package KoNLP version 0.80.2 Index]