concordance_file {KoNLP}	R Documentation

concordance for input text file

Description

returns concordance text for input file.

Usage

concordance_file(filename, pattern, encoding = getOption("encoding"),
  span = 5)

Arguments

filename
file name

pattern
patterns of central words

encoding
filename's encoding

span
how many character will be produced around input pattern

Author(s)

Heewon Jeon

References

Church, K. W. and Mercer, R. L. (1993). Introduction to the special issue on computational linguistics using large corpora. Computational Linguistics, 19(1):1-24.

[Package KoNLP version 0.80.2 Index]