reloadUserDic {KoNLP}	R Documentation

reload dictionaries for specific functions

Description

This function for reloading user dictionary for specific functions, after you have updated user dictionary on KoNLP_dic/current/user_dic.txt.

Usage

reloadUserDic(whichDics)

Arguments

whichDics
character vector which can be "extractNoun", "SimplePos09", "SimplePos22", "SimplePos22"

Examples

## Not run: 
reloadUserDic(c("extractNoun", "SimplePos22"))
## End(Not run) 

[Package KoNLP version 0.80.2 Index]